all artists in playlist print now!
woo! all artists in a playlist are now printing successfully! added some more logic in order to make that happen. Basically, the very first json returned with the first 100 songs of a playlist (spotify set that as the max) has a completely different format than every song json above 100. In order to call songs 100-200, you need to navigate the 0-100 json, and in order to call any song above 200, you need to navigate the (differently formatted) 100-200 json. Whew.